Project Folder Restructuring

+++ MyPocketCal2003/Class Files/UnitConversion.cs
@@ -27,25 +27,25 @@ namespace MyPocketCal2003
 
             //converting from inputUnit to baseUnit....baseUnit act as intermediary
             XmlNode conversionNode;
-            double inputToBaseRatio = 1.0;
+            Decimal inputToBaseRatio = 1.0M;
 
             if (!inputUnit.Equals(baseUnit)) //there is no ratio available for going from A unit to A unit i.e. inputUnit=baseUnit
             {
                 conversionNode = quantityNode.SelectSingleNode("Conversion[FromUnit='" + inputUnit + "' and ToUnit='" + baseUnit + "']");
-                inputToBaseRatio = Convert.ToDouble(conversionNode.LastChild.InnerText); //inputUnit to baseUnit conversion ratio
+                inputToBaseRatio = Convert.ToDecimal(conversionNode.LastChild.InnerText); //inputUnit to baseUnit conversion ratio
             }
 
-            double baseToOutputRatio = 1.0;
+            Decimal baseToOutputRatio = 1.0M;
 
             if (!outputUnit.Equals(baseUnit)) //there is no ratio available for going from A unit to A unit i.e. baseUnit=outputUnit
             {
                 //converting from baseUnit (intermediary) to outputUnit
                 conversionNode = quantityNode.SelectSingleNode("Conversion[FromUnit='" + baseUnit + "' and ToUnit='" + outputUnit + "']");
-                baseToOutputRatio = Convert.ToDouble(conversionNode.LastChild.InnerText); //baseUnit to outputUnit conversion ratio
+                baseToOutputRatio = Convert.ToDecimal(conversionNode.LastChild.InnerText); //baseUnit to outputUnit conversion ratio
             }
 
             //converting input to base unit then converting base unit to output unit
-            return Convert.ToString((Convert.ToDouble(input)*inputToBaseRatio)*baseToOutputRatio);
+            return Convert.ToString((Convert.ToDecimal(input)*inputToBaseRatio)*baseToOutputRatio);
         }
     }
 }
